Fix unary tests

+++ test/cpp/models/xor.cpp
@@ -99,7 +99,7 @@ TEST(models, xor_e5m10) {
         tensor loss {nn::optimizer::mse(y_hat, y)};
         loss.backward();
         if (epoch % 100 == 0) {
-            std::cout << "Epoch: " << epoch << ", Loss: " << loss.to_vector<float>()[0] << std::endl;
+            std::cout << "Epoch: " << epoch << ", Loss: " << static_cast<float>(loss.to_vector<float16>()[0]) << std::endl;
         }
         optimizer.step();
         optimizer.zero_grad();
